requirements.txt
psycopg2-binary
python-dotenv
sentence-transformers
chromadb
fastapi
uvicorn
"pydantic[email]"
bcrypt
ollama